Move ClutchState to lib.
